pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdagent-quant-opt"
version = "0.1.0"
description = "Cost-efficient multi-model routing for automated quantitative strategy development"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "rdagent>=0.3.0",
    "litellm>=1.40.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
backtest = ["pyqlib>=0.9.0"]
analysis = ["pandas>=2.0", "matplotlib>=3.7", "seaborn>=0.12", "jupyter>=1.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py310"
line-length = 100
